% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{print.invasimapr_fit}
\alias{print.invasimapr_fit}
\title{Compact print method for invasimapr_fit}
\usage{
\method{print}{invasimapr_fit}(x, ...)
}
\description{
Internal print method. Not part of the public API.
}
\keyword{internal}
