% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{utils_internal}
\alias{utils_internal}
\alias{new_invasimapr_fit}
\alias{.standardise_df}
\alias{.scale_like}
\alias{.row_z}
\title{Internal utilities}
\description{
Small helper functions used inside the package. Not part of the public API.
}
\keyword{internal}
